+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Upcoming
+### Additions
+* Config option to disable respawn obelisks generation
+
 ### Changes
 * Experience to level up reduced (50 -> 40)
 * Insomnia game rule is now set to false when tiredness is enabled